# openapi.yaml
openapi: 3.0.3
info:
  title: Hexade Case Management API
  description: |
    Modern case management and hearing platform for legal professionals.
    
    ## Features
    - WebRTC video hearings with real-time chat
    - Document management with e-signing capabilities
    - AI-powered document summarization
    - Role-based access control (Client, Lawyer, Judge, Admin)
    - Multi-language support
    
    ## Authentication
    All API endpoints require authentication via JWT tokens stored in secure HTTP-only cookies.
    
    ## Rate Limiting
    API requests are rate limited to 100 requests per 15-minute window per user.
    
    ## Environment Variables Required
    - `DATABASE_URL`: PostgreSQL connection string
    - `JWT_SECRET`: Secret key for JWT token signing
    - `S3_BUCKET`: AWS S3 bucket name for document storage
    - `S3_REGION`: AWS S3 region
    - `S3_ACCESS_KEY_ID`: AWS access key ID
    - `S3_SECRET_ACCESS_KEY`: AWS secret access key
    - `SENDGRID_API_KEY`: SendGrid API key for email notifications
    - `OPENAI_API_KEY`: OpenAI API key for AI summarization (optional)
    - `REDIS_URL`: Redis connection string for background jobs (optional)
  version: 1.0.0
  contact:
    name: Hexade Team
    email: contact@hexade.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://hexade-portal.vercel.app/api
    description: Production server
paths:
...
  /hearings:
    get:
      summary: Get hearings
      tags:
        - Hearings
      security:
        - cookieAuth: []
      parameters:
        - name: caseId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of hearings
          content:
            application/json:
              schema:
                type: object
                properties:
                  hearings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Hearing'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorized
...
  /summarize:
    post:
      summary: Generate AI summary
      tags:
        - AI
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                documentId:
                  type: string
                  description: ID of document to summarize
                text:
                  type: string
                  description: Raw text to summarize
                type:
                  type: string
                  enum: [document, case, hearing]
      responses:
        '200':
          description: Summary generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                  documentId:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Validation error
        '401':
          description: Unauthorized
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth-token
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [CLIENT, LAWYER, JUDGE, ADMIN]
        phone:
          type: string
        avatar:
          type: string
        bio:
          type: string
        language:
          type: string
          default: en
        timezone:
          type: string
          default: UTC
        emailNotifications:
          type: boolean
          default: true
        smsNotifications:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Hearing:
      type: object
      properties:
        id:
          type: string
        caseId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in minutes
        status:
          type: string
          enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
        roomId:
          type: string
        meetingUrl:
          type: string
        agenda:
          type: string
        notes:
          type: string
        recordingUrl:
          type: string
        transcriptUrl:
          type: string
        summary:
          type: string
        judge:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...